fix: restore Windows CLI release build
